--- src/samudra/data/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2026 Samudra Authors
-#
-# SPDX-License-Identifier: Apache-2.0
